[MOV] web: move position_hook up one folder
