+++ addons/mail/static/src/components/popover_view/popover_view.js
@@ -3,7 +3,7 @@
 import { useComponentToModel } from '@mail/component_hooks/use_component_to_model/use_component_to_model';
 import { registerMessagingComponent } from '@mail/utils/messaging_component';
 
-import { usePosition } from '@web/core/position/position_hook';
+import { usePosition } from '@web/core/position_hook';
 
 const { Component } = owl;
